Represent mixture as class, with implemented EMs
Can generate static mixtures (i.e. no interaction with window stuff).
The best way of mixture initialization is defining a number of
components and a distribution low in class constuctor, while other
parameters such as mathematical expectation and etc. generates with some
local "initializator" method (for now it's method from Algorithms module).
Moreover it's possible to get samples out of mixture, that could be used in
testing validity of EM algorithms.
Which by the way is expanded to trio: iterative, adaptive (those are primal
ones) and combined (it is essential). Last one consists of (1) parameters
bank foundation with iterative-EM, (2) selection a sub-bank by best
log-likelihood values, (3) applying adaptive-EM for each element in formed
bank and (4) choosing the best parameters out of it.
All content of Mixture class is documented.

Fix random_seed in tpf.distribution

Yore sampling was random, no matter the fact that random seed was fixed.